+++ source/recipes/DescriptiveAlertRecipe.swift
@@ -1,0 +1,15 @@
+//
+//  DescriptiveAlertRecipe.swift
+//  TVMLKitchen
+//
+//  Created by Stephen Radford on 15/03/2016.
+//  Copyright © 2016 toshi0383. All rights reserved.
+//
+
+public class DescriptiveAlertRecipe: AlertRecipe {
+
+    override var templateFile: String {
+        return "DescriptiveAlertRecipe"
+    }
+
+}