+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Add Modal Support  
   [Stephen Radford](https://github.com/steve228uk)  
   [#26](https://github.com/toshi0383/TVMLKitchen/pull/26)
+* Add Alert Recipes  
+  [Stephen Radford](https://github.com/steve228uk)  
+  [#31](https://github.com/toshi0383/TVMLKitchen/pull/31)
 
 ##### Enhancements
 * Add Cocoapods Podspec